requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Database
mysqlclient==2.2.0

# Image handling
Pillow==10.0.1

# Environment variables
python-decouple==3.8

# Development tools
django-debug-toolbar==4.2.0

# API Documentation
drf-yasg==1.21.7

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.9.1
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring and logging
sentry-sdk==1.38.0

# File handling
django-storages==1.14.2

# Excel export
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF generation
reportlab==4.0.5
weasyprint==60.1
